+++ src/components/MyResume/myResume.js
@@ -1,7 +1,7 @@
 import { ResumeButton, CardContainter } from "./styles";
 import { ReactComponent as Cross } from "../../assets/myPage/Cross.svg";
 import MyResumeCard from "../MyPageCard/MyResumeCard/myResumeCard";
-import MyResumeTestData from "../../api/myResumeTestData";
+import MyResumeTestData from "../../api/testDummyData/myResumeTestData";
 
 const MyResume = ({ setIsNewResume }) => {
   return (
