+++ src/components/MySupport/mySupport.js
@@ -3,7 +3,7 @@ import { Completed, CompletedLabel, CardContainter } from "./styles";
 import { ReactComponent as CheckCircleOff } from "../../assets/myPage/CheckCircleOff.svg";
 import { ReactComponent as CheckCircleOn } from "../../assets/myPage/CheckCircleOn.svg";
 import MySupportCard from "../MyPageCard/MySupportCard/mySupportCard";
-import supportTestData from "../../api/supportTestData";
+import supportTestData from "../../api/testDummyData/supportTestData";
 
 const MySupport = (props) => {
   const [isCompleted, setIsCompleted] = useState(false);
